+++ src/main/java/com/tp3/repository/DocumentRepository.java
@@ -11,5 +11,6 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface DocumentRepository extends JpaRepository<Document, Long> {
+public interface DocumentRepository extends JpaRepository<Document, Integer> {
+
 }
