+++ src/main/java/com/tp3/repository/UserRepository.java
@@ -3,5 +3,5 @@ package com.tp3.repository;
 import com.tp3.model.Utilisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<Utilisateur, Long> {
+public interface UserRepository extends JpaRepository<Utilisateur, Integer> {
 }
